[ci skip] [debug] stashed text-tokenless debug
