+++ lib/cluster.js
@@ -73,6 +73,7 @@ class Cluster {
     * Expands address_clusters with identically-named but unmatched nearby address clusters
     **/
     adoption(cb) {
+	return cb();
         this.pool.query(`
                 BEGIN;
                 ALTER TABLE network_cluster DROP COLUMN IF EXISTS address_text;
@@ -97,6 +98,7 @@ class Cluster {
     * Prune network_cluster assignments where the address_cluster is assigned to more than one
     **/
     prune(cb) {
+	return cb();
 	let that = this;
 	this.pool.connect((err, client, pg_done) => {
             if (err) return cb(err);
