+++ lib/linker.js
@@ -7,6 +7,14 @@ const diacritics = require('diacritics').remove
 module.exports = (street, addresses) => {
     let current;
 
+    let debugId = false;
+    if (process.env.LINKER_DEBUG) {
+	if (parseInt(process.env.LINKER_DEBUG) === street.id) {
+	    debugId = street.id;
+	    console.error(`=== [linker] running for ${street.text} ===`); 
+	}
+    }
+
     for (let addr_it = 0; addr_it < addresses.length; addr_it++) {
         let address = addresses[addr_it];
 
@@ -47,15 +55,22 @@ module.exports = (street, addresses) => {
         }
 
         //Calculate % Match
+	console.error(address.text, street.text);
+	console.error(address.text_tokenless, street.text_tokenless);
         let score = 100 - (((2 * levScore) / (address.text.length + street.text.length)) * 100)
 
         if (!current || current.score < score) {
+	    if (debugId)
+		console.error(`    - new leader: ${address.text}`);
             current = {
                 score: score,
                 address: address
-            }
+            };
         }
     }
+
+
+    console.error(current);
 
     //There is a potential address match and it has a lev. distance of over 40% the same
     if (current && current.address && current.score > 40)  return current.address;
