+++ lib/map.js
@@ -18,6 +18,7 @@ const buffer = require('./buffer');
 const split = require('./split');
 const linesplit = require('split');
 const misc = require('./misc');
+const tokenize = require('./tokenize');
 
 module.exports = function(argv, cb) {
     if (!cb || typeof cb !== 'function') throw new Error('lib/map.js requires a callback parameter');
@@ -71,28 +72,8 @@ module.exports = function(argv, cb) {
 
     if (argv.segment) argv.segment = path.resolve(__dirname, '..',  argv.segment);
 
-    if (argv.tokens) {
-        argv.tokens = argv.tokens.split(',');
-
-        let parsed = [];
-        argv.tokens.forEach((token) => {
-            parsed = parsed.concat(tokens(token, true)); // pull singletons in, too -- ie tokens that are common but have no abbreviation
-        });
-
-        let parsedTokens = {};
-        parsed.forEach((parse) => {
-            parse.sort((a, b) => {
-                return a.length > b.length
-            });
-
-            // we intentionally mark the smallest token as a replacement for itself
-            // this seems silly but it lets us exclude it from text_tokenless in cases where it's pre-abbreviated
-            for (let pi = 0; pi < parse.length; pi++)
-                parsedTokens[parse[pi].toLowerCase()] = parse[0].toLowerCase();
-
-            argv.tokens = parsedTokens;
-        });
-    }
+    if (argv.tokens)
+	argv.tokens = tokenize.createReplacer(argv.tokens.split(','));
 
     const output = fs.createWriteStream(path.resolve(__dirname, '..', argv.output));
