+++ lib/match.js
@@ -67,6 +67,8 @@ function match(id, cb) {
             text_tokenless: res.rows[0].network_text_tokenless
         }, res.rows)
 
+	console.error(address);
+
         if (!address) return cb();
 
         pool.query(`
