+++ lib/tokenize.js
@@ -1,6 +1,9 @@
+const tokens = require('@mapbox/geocoder-abbreviations')
+
 module.exports = {};
 module.exports.main = main;
 module.exports.createGlobalReplacer = createGlobalReplacer;
+module.exports.createReplacer = createReplacer;
 module.exports.replaceToken = replaceToken;
 
 const _ = require('lodash');
@@ -46,6 +49,9 @@ function main(query, replacer, complex) {
             tokens.push(pretokens[i]);
         }
     }
+
+
+    console.error('>>>', replacer);
 
     for (let i = 0; i < tokens.length; i++) {
         if (replacer[tokens[i]])
@@ -99,4 +105,24 @@ function createGlobalReplacer(tokens) {
         replacers.push(entry);
     }
     return replacers;
+}
+
+function createReplacer(languages) {
+    let parsed = [];
+    languages.forEach((token) => {
+	parsed = parsed.concat(tokens(token, true)); // pull singletons in, too -- ie tokens that are common but have no abbreviation
+    });
+
+    let parsedTokens = {};
+    parsed.forEach((parse) => {
+	parse.sort((a, b) => {
+	    return a.length > b.length
+	});
+
+	// we intentionally mark the smallest token as a replacement for itself
+	// this seems silly but it lets us exclude it from text_tokenless in cases where it's pre-abbreviated
+	for (let pi = 0; pi < parse.length; pi++)
+	    parsedTokens[parse[pi].toLowerCase()] = parse[0].toLowerCase();
+    });
+    return parsedTokens;
 }
